bot4mates/requirements.txt
colorama==0.4.6
python-dateutil==2.8.2
Markdown==3.4.1